+++ bht_config.py
@@ -22,6 +22,8 @@ with open(config_file) as f:
 with open(os.path.join(BHT_ROOT_DIR, "VERSION.txt")) as version_file:
     yml_settings["VERSION"] = version_file.read().strip()
 
+yml_settings["BHT_VERSION"] = "1"
+
 if not os.path.isabs(yml_settings["BHT_DATA_DIR"]):
     yml_settings["BHT_DATA_DIR"] = os.path.join(
         BHT_ROOT_DIR, yml_settings["BHT_DATA_DIR"]