+++ tests/unit/test_frontend.py
@@ -8,6 +8,28 @@ from web.main.routes import get_file_from_url, save_to_db
 from web.models import Paper
 
 
+class TestCatalogs:
+    def test_catalog_page(self, test_client):
+        """
+        GIVEN a flask app
+        WHEN the catalog page is requested
+        THEN check that page is returned
+        """
+        response = test_client.get("/catalogs")
+        assert response.status_code == 200
+        assert b"Catalogs by Mission" in response.data
+
+    def test_catalog_page_2(self, test_client, paper_for_test):
+        """
+        GIVEN a flask app and paper_for_test in db
+        WHEN the catalog page is requested
+        THEN check that page contains non_added catalogs
+        """
+        response = test_client.get("/catalogs")
+        assert response.status_code == 200
+        assert b"Catalogs to add:" in response.data
+
+
 class TestFrontUtils:
     @skip_slow_test
     def test_get_file_from_url(self):