+++ web/config.py
@@ -1,6 +1,6 @@
 import os
 
-from bht_config import yml_settings
+from bht_config import yml_settings, BHT_ROOT_DIR
 
 
 class Config(object):
@@ -15,10 +15,6 @@ class Config(object):
         yml_settings["WEB_DB_DIR"], "bht_web.db"
     )
 
-    def __init__(self) -> object:
-        # Get config keys from the settings dictionnary
-        self.__dict__.update(yml_settings)
-
 
 class ProdConfig(Config):
     # Simple production wrapper for Config
@@ -28,7 +24,7 @@ class ProdConfig(Config):
 class DevConfig(Config):
     # Set DEBUG with `flask run --debug` option
     TESTING = True
-    UPLOAD_FOLDER = "dev-upload/"
+    WEB_UPLOAD_DIR = os.path.join(BHT_ROOT_DIR, "dev-upload/")
 
 
 class TestConfig(Config):
@@ -36,6 +32,7 @@ class TestConfig(Config):
     # Set DEBUG with `flask run --debug` option
     # Ignore @login_required decorator
     # LOGIN_DISABLED = True
-    UPLOAD_FOLDER = "test-upload/"
+    WEB_UPLOAD_DIR = os.path.join(BHT_ROOT_DIR, "test-upload/")
     # Set sqlite to in memory for tests
     SQLALCHEMY_DATABASE_URI = "sqlite:///:memory:"
+    BHT_RESOURCES_DIR = os.path.join(BHT_ROOT_DIR, "resources-tests")